pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic-searcher"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic-settings>=2.0",
    "torch>=2.0",
    "open-clip-torch>=2.29",
    "Pillow>=10.0",
    "numpy>=1.26",
    "tokenizers>=0.21",
    "sqlalchemy[asyncio]>=2.0",
    "aiomysql>=0.2",
    "pymysql>=1.1",
    "alembic>=1.14",
    "httpx>=0.28",
    "beautifulsoup4>=4.12",
    "readability-lxml>=0.8",
    "lxml>=5.0",
    "python-dotenv>=1.0",
    "websockets>=14.0",
    "dnspython>=2.8",
    "langdetect>=1.0.9",
    "playwright>=1.40",
    "tldextract>=5.0",
    "qdrant-client>=1.12",
    "symspellpy>=6.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
